+++ src/io_utils.c
@@ -32,7 +32,8 @@ size_t xfread(void *ptr, size_t size, size_t nmemb, FILE *stream)
 {
 	size_t ret = fread(ptr, size, nmemb, stream);
 	if (ret != nmemb)
-		__ERROR("fread, wrong file or file is corrupted nmem expected %d found %d\n", nmemb, ret);
+		__ERROR("fread, wrong file or file is corrupted nmem expected %d found %d\n",
+				(int) nmemb, (int) ret);
 	return ret;
 }
